[DocsUI, Search]: Introduces new loaders for search v2.
Introduces a hierarchy of new loaders. The BaseFileLoader is the base
class from which FolderLoader and SearchLoader are derived. It holds
onto the common objects used by both loaders. It defines the
queryLocation method that tries, with all known users, to query the
given location. It returns the first non-null cursor it gets from the
ContentProviderClient. These loaders are created in the same location as
the old set of loaders, based on the useSearchV2 flag. The CL adds a
very preliminary unit tests for the new loaders.

Test: m DocumentsUIGoogle
Bug: 378590632
Change-Id: Id8da196ef79f9a84dff12f5df5c086c318dd325c
Flag: com.android.documentsui.flags.use_search_v2

Adding test data

Change-Id: I641879d3dea8b0dc3fe84c50a5fbda0ce9b7907a