[DocsUI M3] Update styles for dialogs
This removes the style overrides in preference for the existing styles
that are used by the material 3 library. Currently the border radius
uses the grid_item_radius for setting the corner radius. This is not
strictly true, for now we'll remove this value and defer to the default
one, however, in the future we might want to introduce a new
overlayable.

Bug: 377771488
Test: Deploy with flag on/off
Flag: com.android.documentsui.flags.use_material3
Change-Id: I3292f539eceb8138ebdab8aca139678d6ceabaf5