Show the "Extract to" menu item as an action
It makes it easier to see and access this action when it is presented as
a button instead of being "hidden" in a menu. When files are selected
from a mounted archive, this is pretty much the only sensible action to
select. So, let's make it more visible.

Moreover, the "Extract" icon is already shipped with this app. But it
wasn't used so far.

Bug: 391983717
Flag: com.android.documentsui.flags.use_material3
Test: Manual
Change-Id: I8b8e5c8b5eda4e4391ed1b8882fdd6a5f7e7e591