Revert "Update the file list sort icon"
This reverts commit a71b4b97b16039db9dd0d93f087d468a03383e95.

Reason for revert: This test is failing on all environments. It's not
immediately clear why it's being ran on bots that don't have the flag
enabled. We have a train to board this week, so let's revert the test
for now (the flag is disabled by default) and reland once we've worked
out why it's running on those environments.

Bug: b/393339189, b/377771803
Change-Id: I240dac105fad1df76de54d6fa045ec39df7c7c0d