Support test flag annotation filtering in SortDocumentUiTest
Filtering tests based on flags is a feature that is only available when
running tests cases in Junit 4. Unfortunately all the functional tests
are written in Junit 3. To update to Junit 4 it requires rewriting some
of the base classes, specifically ActivityTest.

The current ActivityTest uses the deprecated
ActivityInstrumentationTestCase2 which was deprecated in favour of
ActivityTestRule. Unforunately this has also since been deprecated in
favour of ActivityScenario. There is quite a bit of churn in the way
tests are setup, so to avoid a large sweeping CL, this introduces a
minimum class required to make the SortDocumentUiTest to work.

Follow up tests will be introduced to migrate the existing functional
tests to Junit 4 to allow for individual tests to be filtered. Once all
tests have been migrated, the file will be renamed back to ActivityTest
and the old (deprecated) version will be removed.

Bug: 393339189, 377771803
Test: atest SortDocumentUiTest#testSortByArrowIcon
Flag: EXEMPT test change
Change-Id: I408d3d7f3f644a9b9d1ce760160aef119d227c3e