Introduce TrampolineActivity to route requests to Photopicker
Currently DocumentsUI has a priority of 100 for the ACTION_GET_CONTENT
intent. When the new Photopicker was released, it took over the
GET_CONTENT intent and routes requests that it can't handle back to
DocumentsUI. On some devices that have DocsUI disabled, Photopicker
ends up in a loop trying to launch DocsUI. To avoid this, perform
the test in DocumentsUI.

This is also a precursor for some follow up work which is going to
migrate DocsUI to pop up via a bottom sheet instead of taking over
the entire window. On devices with WINDOW_MANAGEMENT enabled the
experience of full window takeover is jarring and the bottom sheet
is a more idiomatic UX.

Bug: 377771195
Flag: EXEMPT resource file change
Test: atest com.android.documentsui.TrampolineActivityTest
Change-Id: I9212b3cc52e5dc0b92543061f2a384dacc9c7257