[DocsUI Search] Add a read/write DocsUI search flag.
The search flag is used to control whether the current or the v2
generation of search functionality is used in DocumentsUI. The flag
will be used in both Java code and XML files. This flag is to replace
use_search_v2, which was read-only.

Test: m DocumentsUIGoogle
Bug: 383412640
Flag: com.android.documentsui.flags.use_search_v2_rw
Change-Id: I9706443f43ccc115726d58356ca42795458f8060