Move the selection menu from the ActionBar to NavigationView
In the >600dp layout, the MaterialToolbar has been moved to be
adjacent to the sidebar. Previously the ActionBar was used to replace
the existing toolbar, unfortunately on the new layout this doesn't
work as it spans the full width.

To get around this (and maintain compatibility when the flag is on or
off) we instead inflate the activity menu OR the action mode menu
depending on whether there is a selection. This is a temporary measure
to not have to refactor the click handlers too much. In the final
product the ActionModeController will be removed and the click
handlers consolidated into a single place.

Bug: 383669583
Test: atest DocumentsUIGoogleTests
Flag: com.android.documentsui.flags.use_material3

Change-Id: Ida4675dafbc2096c986067ba4fbe5a695a8e243b