[DocsUI M3] Move drawer burger menu to nav rail
For medium size screen, we show a nav rail on the left side,
if nav rail exists, we don't want to show the drawer burger menu
on the toolbar, instead, we want to show it on the top of the
nav rail.

Check the attached bug for demo video.

Bug: 385652028
Test: m DocumentsUIGoogle && manual inspection
Flag: com.android.documentsui.flags.use_material3
Change-Id: I6dcfbbf99ab16ce1db83fc0002088a2f3e3621f9