Ensure root tests consider the hide_roots_on_desktop flag
Tests in FilesActivityDefaultsUiTest and FilesActivityUiTest that should
only be run when the hide_roots_on_desktop flag is disabled are failing
when it is enabled. Fix this by using @RequiresFlagsDisabled to skip
these tests when the flag is enabled. Also add complement tests for
behaviour when the flag is enabled with @RequiresFlagsEnabled where
appropriate.

Also use ActivityTestJunit4 instead of ActivityTest for the tests to
ensure that @RequiresFlagsDisabled works. See ag/31583362 for more
details.

Bug: 397256734
Test: atest FilesActivityDefaultsUiTest
Test: atest FilesActivityUiTest
Flag: EXEMPT test change
Change-Id: Idb651c0574146167579ef317a8b5426199148a52