Handle case where GET_CONTENT is */* + media extra mime type
The original change (and subsequent roll out of the redirect_get_content
flag) caused a few CTS tests to start failing. This was a result of
a missed edge case where the intent type is */*, however, the intent
has EXTRA_MIME_TYPES set to exclusively media mime types. This should
be forwarded to the photopicker (which will have an escape hatch to
send it back to DocumentsUI).

The logic has been updated along with some general test hygiene where
the photopicker activities were just being started on top of one
another and not being closed. This lead to tests not being
deterministic and randomly failing as well.

Bug: 397672317
Bug: 377771195
Flag: com.android.documentsui.flags.redirect_get_content
Test: atest com.android.documentsui.TrampolineActivityTest
Test: atest android.photopicker.cts.PhotoPickerTest
Test: atest android.photopicker.cts.ActionGetContentOnlyTest

Change-Id: I5e54e1d4ba4219bd38f53232581ed12c93f4b5d5