Add additional debug information for TrampolineActivity
It seems like the TrampolineActivity is failing on a few devices.
Currently there are no logs available, however, when testing it locally
on a bluejay it appears like the PhotopickerGetContentActivity was
disabled. Adding some logs to see if that's the case on the bots.

This also clears up the error messages in the tests which were just
showing up "Assertion failed" which required some small level of leaps
to associate with the test name. Let's just make it obvious in the
failure message what happened and what was expected.

Bug: 398402760
Test: atest com.android.documentsui.picker.TrampolineActivity
Flag: com.android.documentsui.flags.redirect_get_content
Change-Id: I6843d9869bcc3bc0b14d01ac56e5b2585b3f7742