Fix RenameDocumentUiTest#testRenameFile_OkButton
For some reason the onView().perform() logic that is present in the
clickDialogOkButton is not clicking the actual button. This is
leading to the test just dismissing the dialog instead of
actioning it. Instead let's update it to use UiObject2 and perform
the click, this seems more stable across Android versions.

Bug: 394151244
Bug: 384582399
Test: atest RenameDocumentUiTest#testRenameFile_OkButton
Flag: EXEMPT Test fix
Change-Id: I6535a023b66506e363d03b121abb60e6b08c5070