Fix up crash in MenuManager when selection is active
When the menus have been moved from the ActionBar to the NavigationView
it means the menu updating code no longer retrieves the menu from the
BaseActivity. There was some code that was causing a crash on paste in
as it was trying to use the old update path instead. Opt for the new
update method when the flag is enabled instead.

This also adds a regression test that expectedly fails currently but
with the fix will succeed, this also should pass with or without the
flag on.

Fix: 399752575
Bug: 383669583
Flag: com.android.documentsui.flags.use_material3
Test: atest testKeyboard_PasteDocumentWhileSelectionActive
Change-Id: Ic83e1239abee824360a11075ee24c49274f57318