Fix RenameDocumentUitest#testRenameFile_Cancel
For some reason the onView().perform() logic that is present in the
clickDialogCancelButton is not clicking the actual button. This is
leading to the test just leaving the dialog staying on the screen
and the test being unable to find the directoy list. Instead let's
update it to use UiObject2 and perform the click, this seems more
stable across Android versions.

Fix: 385439355
Bug: 394151244
Flag: EXEMPT test fix
Test: atest RenameDocumentUiTest#testRenameFile_Cancel
Change-Id: Ib86d26a2f9c1f1ed24ae9270d4c7b448be9267bc