[DocsUI M3] Uplift focus state for grid items
Previously focus led to a grey overlay over the entire grid item. Now
focus is displayed as a ring around the nameplate. When the item is also
selected, the focus ring wraps the non-transparent nameplate background
with a gap in between.

Since the outer MaterialCardView layer is always transparent, including
during focus, it can be removed and the RelativeLayout can become the
outer "item_root" layer. Set android:defaultFocusHighlightEnabled="false" on the RelativeLayout to
ensure that it remains transparent during focus.

grid_nameplate_background.xml defines the nameplate focus state and to
ensure that the nameplate gets the focus, set the duplicateParentState
property.

See attached bug for screenshots.

Test: m DocumentsUIGoogle && manual inspection
Bug: 400807410
Flag: com.android.documentsui.flags.use_material3
Change-Id: I71af8d10e59b0c27a5b360aca48dd04f336f7d58