Remove (x) in nav rail layout when selection is cleared
In the nav rail layout, the burger menu is shown at the top of the rail
This means when the selection is cleared the (X) should disappear and
not be replaced by the burger menu. It appears like the logic currently
clears the selection but doesn't clear the (X) so it sticks around
despite there not being a selection.

Fix: 400324090
Test: Manual check
Flag: com.android.documentsui.flags.use_material3
Change-Id: Ib462bb4bfae43993319474f627db47a47cf5f43e