Add WindowInsets back on non-freeform window devices
For Material3 the bottom window insets were removed. Unfortunately this
means the navigation gesture swipe at the base of the window overlaps
the picker / saver bottom bar. This is particularly bad in the saver
where the bottom bar is a text box and the gesture navigation interferes
with moving the cursor in the text box.

This also has the benefit of adding the behaviour where the textbox
appears above the soft keyboard. Prior to this if you pressed the
textbox (with material 3 enabled) it would not appear above the soft
keyboard and remain hidden.

Bug: 400337159
Test: Deploy to Pixel 6 and observe bottom bar is above gesture bar
Flag: com.android.documentsui.flags.use_material3
Change-Id: I846ff9718455183bc136fb6ff90adb7c9b590e6d