[DocsUI M3] Remove the break between folders and files
Folders always appear before files and a visual break between them
was introduced in ag/836035. For example, if there is one folder and
several files, the folder appears on the first row and the rest of the
files are displayed below it.

In Material3 we want to display the items continuously, so remove this
break.

See bug for attached screenshots.

Bug: 401108520
Test: m DocumentsUIGoogle && manual inspection
Flag: com.android.documentsui.flags.use_material3
Change-Id: Idd51a10699fe1dad8101c3573afa5d19a04b6c88