Remove fragment_pick_button_background_color and fallback to default
These colors appear to have been introduced to allow for inactive colors
in v31 which is when dynamic colors were introduced. Testing this on
Android 11 it appears like the colors are properly mapped in the
material library and this explicit mapping isn't required.

Fix: 402600004
Test: Deploy to Android 11 device and verify colors are still ok
Flag: com.android.documentsui.flags.use_material3

Change-Id: I4ccae7a4bb1ffc24e87aaeaaf045740b6f3e2cd2