Fix up TrampolineActivityTest and make activity closing more robust
The current TrampolineActivityTest relies on the activity monitor to
close any activities that were found. Unfortunately when 2 activities
are shown OR one flakes, it has a knock on effect to the remainder
tests and causes them to fail. Instead of this, let's get the current
active tasks and close any that are from the documentsui or
photopicker package.

On top of this the extra_mime_types were being added singularly
instead of as a string array. This meant that although it showed
them as appearing to have multiple mime types only one was being
added.

Bug: 377771195
Fix: 401787099
Fix: 401466309
Fix: 400612295
Fix: 401786574
Fix: 398402760
Test: atest com.android.documentsui.TrampolineActivityTest
Flag: com.android.documentsui.flags.redirect_get_content_ro
Change-Id: Ia3f31e5f34da6758930dd3739785fd0cca330209