[DocsUI M3] Fix DirectoryListBot when M3 enabled
The DirectoryListBot implements "selectDocument" for functional test,
relying on the document's icon as a click location. Update the icon
ID based on the changes made in ag/31870100.

The icon_wrapper that we now rely on is also available in list mode,
when M3 is enabled.

Flag: com.android.documentsui.flags.use_material3
Test: com.android.documentsui.*UiTest
Bug: 381962394
Change-Id: I2d835857ac7431b8a68491ba7d05bb27f7567f52