Stop system bar areas from being transparent
On phones the layout is inset to accomodate for the system bars. When
material 3 is enabled, the path bar is moved to the bottom of the
window. When this happens there is a gap created from the inset where
it is made a transparent colour that effectively shows the scrolling
grid / list view. Stop that behaviour and set the default color in M3
to be the background colour.

Fix: 403481518
Test: m DocumentsUIGoogle on bluejay and verify bottom bar
Flag: com.android.documentsui.flags.use_material3
Change-Id: Icb565c80bbb30a6e84486319c1d6914c81bbf03a